TestLangChain/requirement.txt
python-dotenv==1.0.0
langchain==0.0.137
pinecone-client==2.2.1
sentence_transformers
streamlit
streamlit_chat
openai
pinecone-client
unstructured
sentence_transformers